# month_3/lesson_01(mentor)/api/docs/swagger.yaml
definitions:
...
  models.CreateStaff:
    properties:
      branch_id:
        type: string
      name:
        type: string
      password:
        type: string
      staff_type:
        type: string
      tariff_id:
        type: string
      username:
        type: string
    type: object
...
  models.GetAllStaff:
    properties:
      count:
        type: integer
      staffs:
        items:
          $ref: '#/definitions/models.Staff'
        type: array
    type: object
...
  models.Staff:
    properties:
      balance:
        type: number
      branch_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        type: string
      staff_type:
        type: string
      tariff_id:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  response.ErrorResp:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /staff:
    get:
      consumes:
      - application/json
      description: gets all staffs based on limit, page and search by name
      parameters:
      - default: 10
        description: limit
        in: query
        minimum: 1
        name: limit
        type: integer
      - default: 1
        description: page
        in: query
        minimum: 1
        name: page
        type: integer
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetAllStaff'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: GET  ALL STAFFS
      tags:
      - STAFF
    post:
      consumes:
      - application/json
      description: CREATES STAFF BASED ON GIVEN DATA
      parameters:
      - description: staff data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.CreateStaff'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: CREATES STAFF
      tags:
      - STAFF
  /staff/{id}:
    delete:
      consumes:
      - application/json
      description: DELETES STAFF BASED ON ID
      parameters:
      - description: id of staff
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: DELETE STAFF BY ID
      tags:
      - STAFF
    get:
      consumes:
      - application/json
      description: get staff by ID
      parameters:
      - description: Staff ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Staff'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: GET BY ID
      tags:
      - STAFF
    put:
      consumes:
      - application/json
      description: UPDATES STAFF BASED ON GIVEN DATA AND ID
      parameters:
      - description: id of staff
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: staff data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.CreateStaff'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: UPDATE STAFF BY ID
      tags:
      - STAFF
...
swagger: "2.0"